Add auto-merge support to ghb pr
This allows use of --merge or --merge-squash. Ideally you would almost
always use --merge-squash which pulls the commit title / description as
the final squashed commit message, but sometimes you may prefer --merge,
which uses a rebase instead, in the case that you want to maintain
original commit authorship, like in the case of a cherry pick.